% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runMCMC_fn}
\alias{runMCMC_fn}
\title{base function}
\usage{
runMCMC_fn(info, data, constants, code, params, ni, nt, nb)
}
\arguments{
\item{info}{XXX}

\item{data}{XXX}

\item{constants}{XXX}

\item{code}{XXX}

\item{params}{XXX}

\item{ni}{XXX}

\item{nt}{XXX}

\item{nb}{XXX}
}
\value{
result of one chain
}
\description{
\code{runMCMC_fn} base function
}
