% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runMCMC_para}
\alias{runMCMC_para}
\title{Function to run NIBMLE in parallel}
\usage{
runMCMC_para(info, data, constants, code, params, ni, nt, nb)
}
\arguments{
\item{info}{XXX}

\item{data}{XXX}

\item{constants}{XXX}

\item{code}{XXX}

\item{params}{XXX}

\item{ni}{XXX}

\item{nt}{XXX}

\item{nb}{XXX}
}
\value{
out
}
\description{
\code{runMCMC_para} runs any NIMBLE model in parallel using the \code{parallel} package.
}
